+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Added `includes` parameter to `creo2urdf` to include additional yamls.
+
 ## [0.4.7] - 2024-04-09
 - Made `creo2urdf` runnable from terminal